third_party/WebKit/LayoutTests/sensor/idl-AmbientLightSensor.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

test(function() {
  // Test that AmbientLightSensor interface exists
  assert_true('AmbientLightSensor' in window);
  // Test that AmbientLightSensorReading interface exists
  assert_true('AmbientLightSensorReading' in window);
}, 'Test that the AmbientLightSensor and AmbientLightSensorReading interfaces'
   + ' are present.');

</script>
